+++ src/components/Filter.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Checkbox } from 'antd';
+import 'antd/dist/antd.css';
 
 function onChange(checkedValues) {
   console.log('checked = ', checkedValues);
